+++ CHANGELOG.md
@@ -43,6 +43,8 @@
 - docker-compose.yml: postgres:14.10
 - FROM clojure:temurin-21-lein-jammy
 - added `black` in Dockerfile
+
+
 ### Updated
 - home/black-test